add single layer partition demo
We present two partition strategies for a single layer model
(1) partition the model (hidden layer)
(2) partition the batch (input)

We run the foward and backward iteration several times, and also check
that the network weight after training is the same.

We also add timing. Unfortunately, timing for now is not too helpful. It
varies a lot on different runs.